Cast to float32

+++ roll.py
@@ -28,7 +28,7 @@ def roll(filepath, outdir = None):
         pixels = decoder(input_image)
         m, n, chann = tf.shape(pixels).eval()
         # with tf.device('/cpu:0'): # for CPU
-        pix_32 = tf.cast(pixels, tf.int32)
+        pix_32 = tf.cast(pixels, tf.float32)
         with tf.device('/gpu:0'): # for GPU
             rolled_32 = tf.roll(pix_32, shift=[int(m/2), int(n/2)], axis=[0,1])
